Completely insane build script. You should not use this unless you know what you are doing. This might break things in so many ways. Many/most of these flags should be pruned away, but that takes time.
I only post this due to popular demand, not because it is pretty,
safe, or even good settings for everyone.

This will build with LTO enabled, and some really aggressive
optimization flags. BAD IDEA(TM)

If you are lucky, this might give you anything from 0% to 15%
improvement, if you are not lucky, it could give you -30% or even
crash randomly, who knows..